back_end/back_end/celery_service/requirements.txt
psycopg2-binary
celery
redis
python-dotenv
